backend/requirements.txt
cloudinary==1.36.0
Django==4.1
django_environ==0.11.2
djangorestframework==3.14.0
